Move KeyModifiers, ShortcutKeys and LocalizedConsoleKeys types to Eutherion.Shared.csproj, and add missing summaries.
