+++ Eutherion/Shared/Eutherion.Shared.csproj
@@ -72,6 +72,8 @@
     <Compile Include="Text\Json\JsonWhitespace.cs" />
     <Compile Include="Text\TextElement.cs" />
     <Compile Include="Text\TextIndex.cs" />
+    <Compile Include="UIActions\LocalizedConsoleKeys.cs" />
+    <Compile Include="UIActions\ShortcutKeys.cs" />
     <Compile Include="UIActions\UIAction.cs" />
     <Compile Include="UtilityExtensions.cs" />
     <Compile Include="Utils\Box.cs" />
