+++ Eutherion/Win/UIActions/KeyUIActionMapping.cs
@@ -19,118 +19,10 @@
 **********************************************************************************/
 #endregion
 
-using Eutherion.Localization;
 using Eutherion.UIActions;
-using System;
-using System.Collections.Generic;
 
 namespace Eutherion.Win.UIActions
 {
-    [Flags]
-    public enum KeyModifiers
-    {
-        None = 0,
-        Shift = 1,
-        Control = 2,
-        Alt = 4,
-    }
-
-    /// <summary>
-    /// Represents a shortcut key combination for a <see cref="UIAction"/>.
-    /// </summary>
-    public struct ShortcutKeys
-    {
-        public KeyModifiers Modifiers;
-        public ConsoleKey Key;
-
-        public ShortcutKeys(KeyModifiers modifierCode, ConsoleKey key)
-        {
-            Modifiers = modifierCode;
-            Key = key;
-        }
-
-        public ShortcutKeys(ConsoleKey key) : this(KeyModifiers.None, key)
-        {
-        }
-
-        public bool IsEmpty => Key == 0;
-
-        public IEnumerable<LocalizedStringKey> DisplayStringParts()
-        {
-            if (IsEmpty) yield break;
-
-            if (Modifiers.HasFlag(KeyModifiers.Control)) yield return LocalizedConsoleKeys.ConsoleKeyCtrl;
-            if (Modifiers.HasFlag(KeyModifiers.Shift)) yield return LocalizedConsoleKeys.ConsoleKeyShift;
-            if (Modifiers.HasFlag(KeyModifiers.Alt)) yield return LocalizedConsoleKeys.ConsoleKeyAlt;
-
-            if (Key >= ConsoleKey.D0 && Key <= ConsoleKey.D9) yield return LocalizedStringKey.Unlocalizable(Convert.ToString((int)Key - (int)ConsoleKey.D0));
-            else
-            {
-                switch (Key)
-                {
-                    case ConsoleKey.Add:
-                        yield return LocalizedStringKey.Unlocalizable("+");
-                        break;
-                    case ConsoleKey.Subtract:
-                        yield return LocalizedStringKey.Unlocalizable("-");
-                        break;
-                    case ConsoleKey.Multiply:
-                        yield return LocalizedStringKey.Unlocalizable("*");
-                        break;
-                    case ConsoleKey.Divide:
-                        yield return LocalizedStringKey.Unlocalizable("/");
-                        break;
-                    case ConsoleKey.Delete:
-                        yield return LocalizedConsoleKeys.ConsoleKeyDelete;
-                        break;
-                    case ConsoleKey.LeftArrow:
-                        yield return LocalizedConsoleKeys.ConsoleKeyLeftArrow;
-                        break;
-                    case ConsoleKey.RightArrow:
-                        yield return LocalizedConsoleKeys.ConsoleKeyRightArrow;
-                        break;
-                    case ConsoleKey.UpArrow:
-                        yield return LocalizedConsoleKeys.ConsoleKeyUpArrow;
-                        break;
-                    case ConsoleKey.DownArrow:
-                        yield return LocalizedConsoleKeys.ConsoleKeyDownArrow;
-                        break;
-                    case ConsoleKey.Home:
-                        yield return LocalizedConsoleKeys.ConsoleKeyHome;
-                        break;
-                    case ConsoleKey.End:
-                        yield return LocalizedConsoleKeys.ConsoleKeyEnd;
-                        break;
-                    case ConsoleKey.PageUp:
-                        yield return LocalizedConsoleKeys.ConsoleKeyPageUp;
-                        break;
-                    case ConsoleKey.PageDown:
-                        yield return LocalizedConsoleKeys.ConsoleKeyPageDown;
-                        break;
-                    default:
-                        yield return LocalizedStringKey.Unlocalizable(Key.ToString());
-                        break;
-                }
-            }
-        }
-    }
-
-    public static class LocalizedConsoleKeys
-    {
-        public static readonly LocalizedStringKey ConsoleKeyAlt = new LocalizedStringKey(nameof(ConsoleKeyAlt));
-        public static readonly LocalizedStringKey ConsoleKeyCtrl = new LocalizedStringKey(nameof(ConsoleKeyCtrl));
-        public static readonly LocalizedStringKey ConsoleKeyDelete = new LocalizedStringKey(nameof(ConsoleKeyDelete));
-        public static readonly LocalizedStringKey ConsoleKeyDownArrow = new LocalizedStringKey(nameof(ConsoleKeyDownArrow));
-        public static readonly LocalizedStringKey ConsoleKeyEnd = new LocalizedStringKey(nameof(ConsoleKeyEnd));
-        public static readonly LocalizedStringKey ConsoleKeyHome = new LocalizedStringKey(nameof(ConsoleKeyHome));
-        public static readonly LocalizedStringKey ConsoleKeyLeftArrow = new LocalizedStringKey(nameof(ConsoleKeyLeftArrow));
-        public static readonly LocalizedStringKey ConsoleKeyPageDown = new LocalizedStringKey(nameof(ConsoleKeyPageDown));
-        public static readonly LocalizedStringKey ConsoleKeyPageUp = new LocalizedStringKey(nameof(ConsoleKeyPageUp));
-        public static readonly LocalizedStringKey ConsoleKeyRightArrow = new LocalizedStringKey(nameof(ConsoleKeyRightArrow));
-        public static readonly LocalizedStringKey ConsoleKeyShift = new LocalizedStringKey(nameof(ConsoleKeyShift));
-        public static readonly LocalizedStringKey ConsoleKeyUpArrow = new LocalizedStringKey(nameof(ConsoleKeyUpArrow));
-    }
-
     /// <summary>
     /// Represents a mapping between a shortcut key and a <see cref="UIAction"/>.
     /// </summary>
